+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## Version 3.0.1
+
+- [ADD] CIE XYZ Color Space
+- `toXYZAComponents()` method
+
 ## [Version 3.0.0](https://github.com/yannickl/DynamicColor/releases/tag/3.0.0)
 *Released on 2016-06-14.*